+++ hm-dianping/src/main/java/com/hmdp/controller/BlogCommentsController.java
@@ -1,20 +1,28 @@
 package com.hmdp.controller;
 
+import com.hmdp.dto.Result;
+import com.hmdp.entity.BlogComments;
+import com.hmdp.service.IBlogCommentsService;
+import org.springframework.web.bind.annotation.*;
 
-import org.springframework.web.bind.annotation.RequestMapping;
+import javax.annotation.Resource;
 
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * <p>
- *  前端控制器
- * </p>
- *
- * @author 虎哥
- * @since 2021-12-22
- */
 @RestController
-@RequestMapping("/blog-comments")
+@RequestMapping("/blog/comments")
 public class BlogCommentsController {
 
+    @Resource
+    private IBlogCommentsService blogCommentsService;
+
+    @PostMapping("/save")
+    public Result saveBlogComment(@RequestBody BlogComments comment) {
+        return blogCommentsService.saveBlogComment(comment);
+    }
+
+    @GetMapping("/list/{blogId}")
+    public Result queryBlogComments(
+            @PathVariable("blogId") Long blogId,
+            @RequestParam(value = "current", defaultValue = "1") Integer current) {
+        return blogCommentsService.queryBlogComments(blogId, current);
+    }
 }