+++ hm-dianping/src/main/java/com/hmdp/service/IBlogCommentsService.java
@@ -1,16 +1,15 @@
 package com.hmdp.service;
 
-import com.hmdp.entity.BlogComments;
 import com.baomidou.mybatisplus.extension.service.IService;
+import com.hmdp.entity.BlogComments;
 
 /**
  * <p>
  *  服务类
  * </p>
- *
- * @author 虎哥
- * @since 2021-12-22
  */
 public interface IBlogCommentsService extends IService<BlogComments> {
+    com.hmdp.dto.Result saveBlogComment(BlogComments comment);
 
+    com.hmdp.dto.Result queryBlogComments(Long blogId, Integer current);
 }