+++ hm-dianping/src/main/resources/tb_blog_comments.sql
@@ -1,0 +1,14 @@
+CREATE TABLE `tb_blog_comments` (
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
+  `user_id` bigint(20) unsigned NOT NULL COMMENT '用户id',
+  `blog_id` bigint(20) unsigned NOT NULL COMMENT '探店笔记id',
+  `parent_id` bigint(20) unsigned NOT NULL COMMENT '关联的1级评论id，如果是一级评论，则值为0',
+  `answer_id` bigint(20) unsigned NOT NULL COMMENT '回复的评论id',
+  `content` varchar(255) NOT NULL COMMENT '回复的内容',
+  `liked` int(8) unsigned DEFAULT '0' COMMENT '点赞数',
+  `status` tinyint(1) unsigned DEFAULT '0' COMMENT '状态，0：正常，1：被举报，2：禁止查看',
+  `create_time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
+  `update_time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '更新时间',
+  PRIMARY KEY (`id`),
+  KEY `idx_blog_id` (`blog_id`)
+) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;